# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management System API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Filter reports",
                "operationId": "7ae2cb62e2000617a42ef1fd66ed82d5",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "reporter_id",
                        "in": "query",
                        "description": "Filter by reporter ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportCollection"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reports/{reporterId}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get reports by reporter ID",
                "operationId": "88717ca0d7ad6323e5739952347203d4",
                "parameters": [
                    {
                        "name": "reporterId",
                        "in": "path",
                        "description": "Reporter ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Report"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No reports found for this reporter"
                    },
                    "500": {
                        "description": "An error occurred while retrieving the reports"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Report": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "reporter_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "typeOfCategory_id": {
                        "type": "integer"
                    },
                    "location_id": {
                        "type": "integer"
                    },
                    "report_detail_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ReportCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Report"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string"
                            },
                            "last": {
                                "type": "string"
                            },
                            "prev": {
                                "type": "string"
                            },
                            "next": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer"
                            },
                            "from": {
                                "type": "integer"
                            },
                            "last_page": {
                                "type": "integer"
                            },
                            "path": {
                                "type": "string"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "to": {
                                "type": "integer"
                            },
                            "total": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Reports",
            "description": "API Endpoints for Reports"
        }
    ]
}